Fix: refactor analytics from lib to utils
