--- src/lib/analytics/index.js
@@ -1,7 +1,0 @@
-import logPageView from './logPageView';
-import initialize from './initialize';
-
-export {
-  logPageView,
-  initialize
-};
