+++ src/utils/analytics/config.js
@@ -1,4 +1,4 @@
 export default {
-  trackerId: '',
+  trackerId: 'UA-99690204-1',
   debug: true
 };
