+++ src/utils/analytics/index.js
@@ -1,0 +1,7 @@
+import initialize from './initialize';
+import logPageView from './logPageView';
+
+export {
+  initialize,
+  logPageView
+};
