+++ src/utils/analytics/logPageView.js
